chore: Refactor auth module into proposed structure
